Move filter-and-sort-data-view.ts to utils/filter-sort-and-paginate.ts
