+++ packages/dataviews/src/index.ts
@@ -1,7 +1,7 @@
 export { default as DataViews } from './components/dataviews';
 export { default as DataViewsPicker } from './components/dataviews-picker';
 export { default as DataForm } from './components/dataform';
+export { default as filterSortAndPaginate } from './utils/filter-sort-and-paginate';
 export { VIEW_LAYOUTS } from './dataviews-layouts';
-export { filterSortAndPaginate } from './filter-and-sort-data-view';
 export type * from './types';
 export { isItemValid } from './validation';
